Time: 4 ms (44.88%), Space: 23.4 MB (56.49%) - LeetHub

+++ 0155-min-stack/0155-min-stack.cpp
@@ -1,7 +1,7 @@
 class MinStack {
 public:
 
-    stack<int>st;
+    stack<long long>st;
     long long mini = INT_MAX;
     MinStack() {
     }
